refactor enhancer to process non-pod objects
