+++ enhancers_test.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 )
 
-func TestRunPodEnhancer(t *testing.T) {
+func TestRunEnhancers(t *testing.T) {
 
 	// Create empty context
 	ctx := context.Background()
@@ -53,9 +53,9 @@ func TestRunPodEnhancer(t *testing.T) {
 	// Add event message
 	event.Message = "This event is for TestRunPodEnhancer"
 	// Call pod enhancer to modify scope and event
-	err = runPodEnhancer(ctx, podObj, scope, event)
+	err = runEnhancers(ctx, nil, "Pod", podObj, scope, event)
 	if err != nil {
-		t.Errorf("pod enhancer returned an error: %v", err)
+		t.Errorf("runEnhancers returned an error: %v", err)
 	}
 
 	// Apply the scope to the event
